+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Unreleased
 ### Bug fix
 - more user-friendly error messages upon missing directory/permissions
 - fix handling of `full_csm` argument during update
+- have motifs fail validation if they don't resemble a connected graph
 
 ### Breaking changes
 - switch to Java 17